+++ src/main/java/pretzel/dreamketcherbe/common/config/RedisConfig.java
@@ -7,7 +7,7 @@ import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
-import pretzel.dreamketcherbe.auth.entity.Token;
+import pretzel.dreamketcherbe.domain.auth.entity.Token;
 
 @Configuration
 @EnableRedisRepositories
