+++ src/main/java/pretzel/dreamketcherbe/common/config/WebConfig.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import pretzel.dreamketcherbe.auth.config.AuthArgumentResolver;
-import pretzel.dreamketcherbe.auth.config.AuthInterceptor;
+import pretzel.dreamketcherbe.domain.auth.config.AuthArgumentResolver;
+import pretzel.dreamketcherbe.domain.auth.config.AuthInterceptor;
 
 import java.util.List;
