+++ src/main/java/pretzel/dreamketcherbe/domain/auth/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package pretzel.dreamketcherbe.auth.dto;
+package pretzel.dreamketcherbe.domain.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
