+++ src/main/java/pretzel/dreamketcherbe/domain/auth/entity/Token.java
@@ -1,4 +1,4 @@
-package pretzel.dreamketcherbe.auth.entity;
+package pretzel.dreamketcherbe.domain.auth.entity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
