+++ src/main/java/pretzel/dreamketcherbe/domain/auth/google/dto/GoogleToken.java
@@ -1,4 +1,4 @@
-package pretzel.dreamketcherbe.auth.google.dto;
+package pretzel.dreamketcherbe.domain.auth.google.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
