+++ src/main/java/pretzel/dreamketcherbe/domain/auth/jwt/JwtProvider.java
@@ -1,14 +1,13 @@
-package pretzel.dreamketcherbe.auth.jwt;
+package pretzel.dreamketcherbe.domain.auth.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import io.jsonwebtoken.security.Keys;
-import pretzel.dreamketcherbe.auth.repository.TokenProvider;
+import pretzel.dreamketcherbe.domain.auth.repository.TokenProvider;
 
 import javax.crypto.SecretKey;
 import java.nio.charset.StandardCharsets;
